Use https for the live site's import endpoint
PopIt sets the cookie in the UI to have a secure flag, which means it
will only be served over https connections. So although the API works
over plain http, the cookies weren't being sent correctly and the API
was returning a 401.

This change hardcodes the live site url, which isn't ideal but it gets
the job done.